Move out videowriter from stream class/add videoreader class
